@@ -1,105 +1,67 @@
 {
-  "name": "case.api",
+  "name": "simulator-ir",
   "version": "1.0.0",
+  "description": "Backend do simulador de declaracao de imposto de renda.",
   "main": "index.js",
-  "repository": "git@github.com:DiegoCiara/case.api.git",
+  "author": "Diego Ciara",
   "license": "MIT",
-  "engines": {
-    "node": ">=20.11.1"
+  "resolutions": {
+    "typeorm": "0.2.37",
+    "pg": "8.7.1"
   },
   "scripts": {
     "dev": "ts-node-dev --files -r tsconfig-paths/register --inspect=0.0.0.0:7007 --transpile-only --exit-child src/server.ts ",
     "start": "node ./build/src/server.js",
     "typeorm": "ts-node-dev ./node_modules/typeorm/cli.js -f ./ormconfig.ts",
     "test": "NODE_ENV=test jest --detectOpenHandles",
-    "build": "npx tsc && tsc-alias",
-    "initialize": "source ./initialize.sh",
-    "up": "./up.sh"
+    "build": "npx tsc && tsc-alias"
   },
   "dependencies": {
-    "@babel/preset-typescript": "^7.24.1",
-    "@ffmpeg-installer/ffmpeg": "^1.1.0",
-    "@google/generative-ai": "^0.9.0",
-    "@types/fluent-ffmpeg": "^2.1.26",
-    "@types/multer": "^1.4.11",
+    "@types/amqplib": "^0.10.6",
+    "@types/nodemailer": "^6.4.17",
     "@types/nodemailer-express-handlebars": "^4.0.5",
     "@types/qrcode": "^1.5.5",
     "@types/speakeasy": "^2.0.10",
-    "amqplib": "^0.10.4",
-    "aws-sdk": "^2.1669.0",
-    "axios": "^1.6.7",
+    "@types/swagger-jsdoc": "^6.0.4",
+    "@types/swagger-ui-express": "^4.1.7",
+    "amqplib": "^0.10.5",
     "bcryptjs": "^2.4.3",
-    "body-parser": "^1.20.2",
-    "chance": "^1.1.8",
     "cors": "^2.8.5",
-    "crypto": "^1.0.1",
-    "crypto-js": "^4.2.0",
-    "csv-stringify": "^6.4.5",
-    "currency.js": "^2.0.4",
-    "date-fns": "^3.6.0",
-    "dotenv": "^10.0.0",
+    "dotenv": "^16.4.7",
     "express": "^4.17.1",
-    "express-handlebars": "^7.1.2",
-    "ffprobe-static": "^3.1.0",
-    "fluent-ffmpeg": "^2.1.3",
-    "fs": "^0.0.1-security",
-    "image-size": "^1.1.1",
-    "inquirer": "^9.2.15",
-    "jimp": "^1.6.0",
-    "jsonwebtoken": "^8.5.1",
-    "media-converter": "^1.1.7",
-    "mediainfo-wrapper": "^1.2.0",
-    "moment": "^2.30.1",
-    "moment-timezone": "^0.5.45",
-    "multer": "^1.4.5-lts.1",
-    "music-metadata": "^10.5.0",
-    "newrelic": "^11.22.0",
-    "nodemailer": "^6.9.13",
-    "nodemailer-express-handlebars": "^6.1.2",
-    "openai": "^4.47.1",
-    "path": "^0.12.7",
+    "express-handlebars": "^8.0.1",
+    "jsonwebtoken": "^9.0.2",
+    "nodemailer": "^6.9.16",
+    "nodemailer-express-handlebars": "^7.0.0",
+    "openai": "^4.80.0",
     "pg": "^8.7.1",
-    "probe-image-size": "^7.2.3",
-    "qrcode": "^1.5.3",
-    "redis": "^4.7.0",
-    "socket.io": "^4.7.4",
-    "socket.io-client": "^4.7.5",
+    "qrcode": "^1.5.4",
+    "socket.io": "^4.8.1",
     "speakeasy": "^2.0.0",
     "stripe": "^17.5.0",
-    "supertest": "^6.1.6",
+    "swagger-jsdoc": "^6.2.8",
+    "swagger-ui-express": "^5.0.1",
     "tsc-alias": "^1.8.8",
-    "typeorm": "^0.2.37",
+    "typeorm": "0.2.37",
     "typeorm-express-query-builder": "^1.5.0",
-    "uuid": "^8.3.2",
-    "wrangler": "^3.69.1",
+    "uuid": "^11.0.5",
     "yarn": "^1.22.11"
   },
   "devDependencies": {
-    "@babel/cli": "^7.24.5",
-    "@babel/core": "^7.24.5",
-    "@babel/preset-env": "^7.24.5",
-    "@types/amqplib": "^0.10.5",
     "@types/bcryptjs": "^2.4.2",
-    "@types/chance": "^1.1.3",
     "@types/cors": "^2.8.12",
-    "@types/express": "^4.17.13",
-    "@types/jest": "^27.0.2",
-    "@types/jsonwebtoken": "^8.5.5",
-    "@types/mongoose": "^5.11.97",
-    "@types/node": "^20.12.12",
-    "@types/nodemailer": "^6.4.15",
-    "@types/redis": "^4.0.11",
-    "@types/supertest": "^2.0.11",
-    "@types/uuid": "^8.3.1",
-    "@typescript-eslint/eslint-plugin": "^4.29.2",
-    "@typescript-eslint/parser": "^4.29.2",
-    "eslint": "^7.32.0",
+    "@types/express": "^5.0.0",
+    "@types/jest": "^29.5.14",
+    "@types/jsonwebtoken": "^9.0.7",
+    "@typescript-eslint/eslint-plugin": "^8.20.0",
+    "@typescript-eslint/parser": "^8.20.0",
+    "eslint": "^9.18.0",
     "jest": "^29.7.0",
     "jest-module-name-mapper": "^0.1.5",
     "ts-jest": "^29.1.4",
     "ts-node": "^10.9.2",
-    "ts-node-dev": "^1.1.8",
-    "tsconfig-paths": "^3.10.1",
-    "typescript": "^5.4.5"
+    "ts-node-dev": "^2.0.0",
+    "tsconfig-paths": "^4.2.0",
+    "typescript": "^5.7.3"
   }
 }
