@@ -1,125 +1,277 @@
 import User from '@entities/User';
-import sendMail from '@src/services/sendEmail';
-import crypto from 'crypto';
 import { Request, Response } from 'express';
-import jwt from 'jsonwebtoken';
 import bcrypt from 'bcryptjs';
-import Access from '@entities/Access';
 import dotenv from 'dotenv';
+import { generateToken } from '@utils/functions/generateToken';
+import speakeasy from 'speakeasy';
+import qrcode from 'qrcode';
+import emailValidator from '@utils/functions/emailValidator';
+import jwt from 'jsonwebtoken';
+
 dotenv.config();
 
 interface UserInterface {
-  name?: string;
-  role?: string;
-  token?: string;
-  picture?: string;
+  name: string;
   email: string;
+  token: string;
   password: string;
-  client: string;
+  secret?: string;
+  tempToken?: string
 }
-
-function generateToken(params = {}) {
-  return jwt.sign(params, `${process.env.SECRET}`, { expiresIn: 84600 });
-}
+/**
+ * @swagger
+ * tags:
+ *   name: Autenticação
+ *   description: Operações relativas à autenticação
+ */
 
 class AuthController {
-  public async authenticate(req: Request, res: Response): Promise<Response> {
+  /**
+   * @swagger
+   * /auth/:
+   *   post:
+   *     summary: Autentica um usuário
+   *     tags: [Autenticação]
+   *     requestBody:
+   *       required: true
+   *       content:
+   *         application/json:
+   *           schema:
+   *             type: object
+   *             properties:
+   *               email:
+   *                 type: string
+   *               password:
+   *                 type: string
+   *     responses:
+   *       200:
+   *         description: Autenticação bem-sucedida
+   *         content:
+   *           application/json:
+   *             schema:
+   *               type: object
+   *               properties:
+   *                 id:
+   *                   type: string
+   *                 email:
+   *                   type: string
+   *                 name:
+   *                   type: string
+   *                 has_configured:
+   *                   type: boolean
+   *       400:
+   *         description: Valores inválidos para o usuário
+   *       404:
+   *         description: Usuário não encontrado
+   *       401:
+   *         description: Senha inválida
+   *       500:
+   *         description: Erro interno na autenticação
+   */
+  public async authenticate(req: Request, res: Response): Promise<void> {
     try {
       const { email, password }: UserInterface = req.body;
-      if (!email || !password) return res.status(400).json({ message: 'Invalid values for User' });
 
-      const user = await User.findOne({
-        where: { email: email },
-        select: ['id', 'email', 'name', 'token_reset_password', 'password_hash', 'picture'],
+      if (!email || !password) {
+        res.status(400).json({ message: 'Valores inválidos para o usuário.' });
+        return;
+      }
+
+      const user = await User.findOne({ where: { email } });
+
+      if (!user) {
+        res.status(404).json({ message: 'Usuário não encontrado.' });
+        return;
+      }
+
+      if (!(await bcrypt.compare(password, user.password_hash))) {
+        res.status(401).json({ message: 'Senha inválida.' });
+        return;
+      }
+
+      const jwtSecret = process.env.SECRET;
+      // Gerar um token de autenticação  temporário que expira em 5 minutos
+      const tempToken = jwt.sign({ email: user.email }, jwtSecret!, {
+        expiresIn: '5m',
       });
 
-      if (!user) return res.status(404).json({ message: 'E-mail inválido!' });
+      await User.update(user.id, { token_auth_secret: tempToken });
 
-      if (!(await bcrypt.compare(password, user.password_hash))) return res.status(404).json({ message: 'Senha inválida!' });
-
-      const accesses = await Access.find({ where: { user: user }, relations: ['workspace', 'workspace.tokens'] });
-
-      const workspacesWorkspace = await Promise.all(
-        accesses.map(async (access) => {
-          const workspace = access.workspace;
-          return {
-            id: workspace.id,
-            role: access.role,
-          };
-        })
-      );
-
-      const flattenedAssistants = workspacesWorkspace.flat();
-
-      return res.status(200).json({
+      res.status(200).json({
         id: user.id,
         email: user.email,
         name: user.name,
-        picture: user.picture,
-        token: generateToken({ id: user.id }),
-        token_reset_password: user.token_reset_password,
-        workspaces: flattenedAssistants,
+        has_configured_2fa: user.has_configured_2fa,
       });
     } catch (error) {
       console.error(error);
-      return res.status(500).json({ error: 'Authenticate failed, try again' });
+      res.status(500).json({ error: 'Erro interno na autenticação.' });
     }
   }
-  public async forgotPassword(req: Request, res: Response): Promise<Response> {
+
+  /**
+   * @swagger
+   * /auth/2fa/{email}:
+   *   get:
+   *     summary: Gera o QR Code para configuração de autenticação de dois fatores
+   *     tags: [Autenticação]
+   *     parameters:
+   *       - in: path
+   *         name: email
+   *         required: true
+   *         description: Email do usuário
+   *         schema:
+   *           type: string
+   *     responses:
+   *       200:
+   *         description: QR Code gerado com sucesso
+   *         content:
+   *           application/json:
+   *             schema:
+   *               type: string
+   *       404:
+   *         description: Usuário não encontrado
+   *       500:
+   *         description: Erro ao gerar o QR Code
+   */
+  public async get2FaQrCode(req: Request, res: Response): Promise<void> {
     try {
-      const { email } = req.body;
+      const { email } = req.params;
 
-      if (!email) return res.status(400).json({ message: 'Invalid values for forgot password' });
+      const user = await User.findOne({ where: { email } });
 
-      const user = await User.findOne({ email });
+      if (!user) {
+        res.status(404).json({
+          message: 'Usuário não encontrado.',
+        });
+        return;
+      }
 
-      if (!user) return res.status(404).json({ message: 'No user found' });
+      if (!user.token_auth_secret) {
+        res.status(404).json({
+          message: 'Token de autenticação de 2 fatores não encontrado, faça o login e tente novamente.',
+        });
+        return;
+      }
+      const jwtSecret = process.env.SECRET;
+      // Verificar o token temporário
+      try {
+        jwt.verify(user.token_auth_secret, jwtSecret!);
+      } catch (err) {
+        res.status(401).json({ message: 'Token de autenticação de 2 fatores expirado ou inválido, faça o login novamente.' });
+        return;
+      }
 
-      const token = crypto.randomBytes(20).toString('hex'); // token que será enviado via email.
-
-      const now = new Date();
-      now.setHours(now.getHours() + 1);
-
-      await User.update(user.id, {
-        token_reset_password: token,
-        reset_password_expires: now,
+      const secret = speakeasy.otpauthURL({
+        secret: user.secret,
+        label: `IR Simulator: ${user.email}`,
+        encoding: 'base32',
       });
 
-      const client = process.env.CLIENT_CONNECTION;
+      const data = await qrcode.toDataURL(secret!);
 
-      // Envie o email e aguarde a conclusão antes de enviar a resposta
-      await sendMail('forgotPassword.html', 'contato', 'Recuperação de Senha', { client, name: user.name, token, email: user.email });
-
-      // Envie a resposta após o envio do email
-      return res.status(200).json({ message: 'Password reset sent' });
+      res.status(200).json({ qr_code: data, message: 'Este QR code expira em 5 minutos' });
     } catch (error) {
       console.error(error);
-      return res.status(500).json({ error: 'Forgot password failed, try again' });
+      res.status(500).json({ error: 'Erro ao gerar o QR Code.' });
     }
   }
+  /**
+   * @swagger
+   * /auth/2fa/verify:
+   *   post:
+   *     summary: Verifica o token da autenticação de dois fatores e retorna o token da sessão (Aqui obtém o token de autenticação da sessão)
+   *     tags: [Autenticação]
+   *     requestBody:
+   *       required: true
+   *       content:
+   *         application/json:
+   *           schema:
+   *             type: object
+   *             properties:
+   *               email:
+   *                 type: string
+   *               secret:
+   *                 type: string
+   *     responses:
+   *       200:
+   *         description: Verificação bem-sucedida
+   *         content:
+   *           application/json:
+   *             schema:
+   *               type: object
+   *               properties:
+   *                 user:
+   *                   type: object
+   *                 token:
+   *                   type: string
+   *       400:
+   *         description: Valores inválidos para verificação
+   *       404:
+   *         description: Usuário não encontrado
+   *       401:
+   *         description: Falha na verificação, tente novamente
+   *       500:
+   *         description: Erro interno na verificação
+   */
 
-  public async resetPassword(req: Request, res: Response): Promise<Response> {
+  public async verifySecret(req: Request, res: Response): Promise<void> {
     try {
-      const { email, password, token }: UserInterface = req.body;
+      const { email, secret }: UserInterface = req.body;
 
-      if (!email || !password || !token) return res.status(400).json({ message: 'Invalid values for User reset password' });
+      if (!email || !emailValidator(email) || !secret) {
+        res
+          .status(400)
+          .json({ message: 'Valores inválidos para verificação.' });
+        return;
+      }
 
-      const user = await User.findOne({ email });
+      const user = await User.findOne({ where: { email } });
 
-      if (!user) return res.status(404).json({ message: 'Invalid values for User reset password' });
+      if (!user) {
+        res.status(404).json({
+          message: 'Usuário não encontrado.',
+        });
+        return;
+      }
 
-      if (token !== user.token_reset_password) return res.status(400).json({ message: 'Token is invalid' });
+      const jwtSecret = process.env.SECRET;
+      // Verificar o token temporário
+      try {
+        jwt.verify(user.token_auth_secret, jwtSecret!);
+      } catch (err) {
+        res.status(401).json({ message: 'Token de autenticação de 2 fatores expirado ou inválido, faça o login novamente.' });
+        return;
+      }
 
-      const now = new Date();
-      if (now > user.reset_password_expires) return res.status(400).json({ message: 'Token expired' });
 
-      const password_hash = await bcrypt.hash(password, 10);
+      const verified = speakeasy.totp.verify({
+        secret: user.secret,
+        encoding: 'base32',
+        token: secret,
+        window: 1,
+      });
 
-      await User.update(user.id, { password_hash, reset_password_expires: undefined, token_reset_password: undefined });
-
-      return res.status(200).json();
+      if (verified) {
+        const token = generateToken({ id: user.id });
+        await User.update(user.id, { has_configured_2fa: true, token_auth_secret: undefined });
+        res.status(200).json({
+          user: {
+            id: user.id,
+            name: user.name,
+            email: user.email,
+            has_configured: user.has_configured_2fa,
+          },
+          token: token,
+        });
+      } else {
+        res
+          .status(401)
+          .json({ message: 'Código inválido, verifique o código informado.' });
+      }
     } catch (error) {
-      return res.status(400).json({ error: 'Cannot reset password, try again' });
+      console.error(error);
+      res.status(500).json({ error: 'Erro interno na verificação.' });
     }
   }
 }
