@@ -1,13 +1,10 @@
-import AuthController from '../app/controllers/AuthController';
-import { ensureAuthenticated } from '../app/middlewares/ensureAuthenticated';
+import AuthController from '@controllers/AuthController';
 import Router from 'express';
 
 const routes = Router();
 
-routes.post('/authenticate', AuthController.authenticate);
-routes.post('/forgot-password', AuthController.forgotPassword);
-routes.put('/reset-password', AuthController.resetPassword);
-
-routes.get('/faw1efawe3f14aw8es3v6awer51xx3/check', ensureAuthenticated, (req,res) => res.status(200).send());
+routes.post('/', AuthController.authenticate);
+routes.get('/2fa/:email', AuthController.get2FaQrCode);
+routes.post('/2fa/verify', AuthController.verifySecret);
 
 export default routes;
